Introduce RabbitMQ multiple producers
